Added mock asserter since same assert mocks and stubs are being used in every test.
