+++ tests/unit-test/tests/test_timer.cpp
@@ -19,9 +19,11 @@
 #include <cstdint>
 
 /* Files under test. */
-#include <ecu/asserter.h>
 #include <ecu/timer.h>
 #include <ecu/interface/itimer.h>
+
+/* Mocks */
+#include <mocks/mock_asserter.hpp>
 
 /* CppUTest. */
 #include <CppUTestExt/MockSupport.h>
@@ -32,50 +34,6 @@
 /*---------------------------------------------------------------------------------------------------------------------------*/
 /*--------------------------------------------------- STUB AND MOCK DECLARATIONS --------------------------------------------*/
 /*---------------------------------------------------------------------------------------------------------------------------*/
-
-class AssertException
-{
-};
-
-
-/**
- * @brief Use when ECU library function calls within the unit test could 
- * cause assert to fire but you wish to ignore it. This throws an exception
- * so control immediately returns to the caller (the unit test). This 
- * prevents library code from running under an assert condition, leading to
- * undefined behavior.
- */
-struct AssertCallOk : public ecu_assert_functor 
-{
-    static void assert_handler(struct ecu_assert_functor *me, const char *file, int line)
-    {
-        (void)me;
-        (void)file;
-        (void)line;
-        throw AssertException();
-    }
-};
-
-
-/**
- * @brief Use when ECU library function calls within the unit test should
- * never fire an assert. Calls a mock which will cause the test to fail
- * due to an unexpected call. Also throws an exception so control immediately
- * returns to the caller (the unit test). This prevents library code from 
- * running under an assert condition, leading to undefined behavior.
- */
-struct AssertCallFail : public ecu_assert_functor 
-{
-    static void assert_handler(struct ecu_assert_functor *me, const char *file, int line)
-    {
-        (void)me;
-        (void)file;
-        (void)line;
-        mock().actualCall("AssertMock::assert_handler");
-        throw AssertException();
-    }
-};
-
 
 template<typename T>
 concept valid_tick_width = (sizeof(T) <= sizeof(ecu_max_tick_size_t));
